+++ cartridge.lua
@@ -46,6 +46,7 @@ local lua_api_topology = require('cartridge.lua-api.topology')
 local lua_api_failover = require('cartridge.lua-api.failover')
 local lua_api_vshard = require('cartridge.lua-api.vshard')
 local lua_api_deprecated = require('cartridge.lua-api.deprecated')
+local lua_api_boxinfo = require('cartridge.lua-api.boxinfo')
 
 local ConsoleListenError = errors.new_class('ConsoleListenError')
 local CartridgeCfgError = errors.new_class('CartridgeCfgError')
@@ -673,6 +674,7 @@ local function cfg(opts, box_opts)
         end
 
         graphql.init(httpd, {prefix = opts.webui_prefix})
+        lua_api_boxinfo.set_webui_prefix(opts.webui_prefix)
 
         if opts.webui_enabled then
             local ok, err = HttpInitError:pcall(webui.init, httpd, {