+++ cartridge/lua-api/boxinfo.lua
@@ -8,6 +8,15 @@ local errors = require('errors')
 
 local pool = require('cartridge.pool')
 local confapplier = require('cartridge.confapplier')
+local vars = require('cartridge.vars').new('cartridge.boxinfo')
+
+vars:new('webui_prefix', nil)
+vars:new('http_host', nil)
+vars:new('http_port', nil)
+
+local function set_webui_prefix(prefix)
+    vars.webui_prefix = prefix
+end
 
 --- Retrieve `box.cfg` and `box.info` of a remote server.
 -- @function get_info
@@ -85,6 +94,20 @@ local function get_info(uri)
             storage_info = box.NULL
         end
 
+        if vars.http_port == nil then
+            local httpd = package.loaded.cartridge.service_get('httpd')
+            local srv_name = {}
+
+            if httpd ~= nil then
+                srv_name = httpd.tcp_server:name()
+            else
+                vars.webui_prefix = nil
+            end
+
+            vars.http_host = srv_name.host
+            vars.http_port = srv_name.port
+        end
+
         local ret = {
             general = {
                 version = box_info.version,
@@ -98,6 +121,9 @@ local function get_info(uri)
                 wal_dir = box_cfg.wal_dir,
                 worker_pool_threads = box_cfg.worker_pool_threads,
                 listen = box_cfg.listen and tostring(box_cfg.listen),
+                http_host = vars.http_host,
+                http_port = vars.http_port,
+                webui_prefix = vars.webui_prefix,
                 ro = box_info.ro,
             },
             storage = {
@@ -191,4 +217,6 @@ _G.__cluster_admin_get_info = get_info
 
 return {
     get_info = get_info,
+
+    set_webui_prefix = set_webui_prefix,
 }