+++ cartridge/webui/gql-boxinfo.lua
@@ -115,6 +115,18 @@ local boxinfo_schema = {
                         kind = gql_types.string,
                         description = 'The binary protocol URI',
                     },
+                    http_port = {
+                        kind = gql_types.int,
+                        description = 'HTTP port',
+                    },
+                    http_host = {
+                        kind = gql_types.string,
+                        description = 'HTTP host',
+                    },
+                    webui_prefix = {
+                        kind = gql_types.string,
+                        description = 'HTTP webui prefix',
+                    },
                     ro = {
                         kind = gql_types.boolean.nonNull,
                         description = 'Current read-only state',